Fix types in C configuration unit test.

+++ uctest/cfg.c
@@ -18,12 +18,12 @@ Test(ConfigFileTest, MethodEmptyConstructor)
     ptpmgmt_cfg f = ptpmgmt_cfg_alloc();
     cr_expect(not(zero(ptr, f)));
     // Check default values
-    cr_expect(eq(i16, f->transportSpecific(f, NULL), 0));
-    cr_expect(eq(i16, f->domainNumber(f, NULL), 0));
-    cr_expect(eq(i16, f->udp6_scope(f, NULL), 0xe));
-    cr_expect(eq(i16, f->udp_ttl(f, NULL), 1));
-    cr_expect(eq(i16, f->socket_priority(f, NULL), 0));
-    cr_expect(eq(i16, f->network_transport(f, NULL), '4'));
+    cr_expect(eq(i8, f->transportSpecific(f, NULL), 0));
+    cr_expect(eq(i8, f->domainNumber(f, NULL), 0));
+    cr_expect(eq(i8, f->udp6_scope(f, NULL), 0xe));
+    cr_expect(eq(i8, f->udp_ttl(f, NULL), 1));
+    cr_expect(eq(i8, f->socket_priority(f, NULL), 0));
+    cr_expect(eq(i8, f->network_transport(f, NULL), '4'));
     cr_expect(eq(str, (char *)f->uds_address(f, NULL), "/var/run/ptp4l"));
     {
         size_t l;
@@ -44,19 +44,19 @@ Test(ConfigFileTest, MethodEmptyConstructor)
 }
 
 // Tests read configuration file
-// i16 read_cfg(ptpmgmt_cfg _this, const char *file)
+// bool read_cfg(ptpmgmt_cfg _this, const char *file)
 Test(ConfigFileTest, MethodReadConfiguration)
 {
     ptpmgmt_cfg f = ptpmgmt_cfg_alloc();
     cr_expect(not(zero(ptr, f)));
     cr_expect(f->read_cfg(f, "utest/testing.cfg"));
     // Check file configuration
-    cr_expect(eq(i16, f->transportSpecific(f, NULL), 7));
-    cr_expect(eq(i16, f->domainNumber(f, NULL), 5));
-    cr_expect(eq(i16, f->udp6_scope(f, NULL), 0xd));
-    cr_expect(eq(i16, f->udp_ttl(f, NULL), 3));
-    cr_expect(eq(i16, f->socket_priority(f, NULL), 11));
-    cr_expect(eq(i16, f->network_transport(f, NULL), '6'));
+    cr_expect(eq(i8, f->transportSpecific(f, NULL), 7));
+    cr_expect(eq(i8, f->domainNumber(f, NULL), 5));
+    cr_expect(eq(i8, f->udp6_scope(f, NULL), 0xd));
+    cr_expect(eq(i8, f->udp_ttl(f, NULL), 3));
+    cr_expect(eq(i8, f->socket_priority(f, NULL), 11));
+    cr_expect(eq(i8, f->network_transport(f, NULL), '6'));
     cr_expect(eq(str, (char *)f->uds_address(f, NULL), "/var/run/dummy"));
     {
         size_t l;
@@ -77,74 +77,74 @@ Test(ConfigFileTest, MethodReadConfiguration)
 }
 
 // Tests get transport specific parameter
-// i16 transportSpecific(ptpmgmt_cfg _this, const char *section)
+// uint8_t transportSpecific(ptpmgmt_cfg _this, const char *section)
 Test(ConfigFileTest, MethodTransportSpecific)
 {
     ptpmgmt_cfg f = ptpmgmt_cfg_alloc();
     cr_expect(not(zero(ptr, f)));
     cr_expect(f->read_cfg(f, "utest/testing.cfg"));
-    cr_expect(eq(i16, f->transportSpecific(f, "dumm"), 9));
-    cr_expect(eq(i16, f->transportSpecific(f, "non"), 7));
+    cr_expect(eq(i8, f->transportSpecific(f, "dumm"), 9));
+    cr_expect(eq(i8, f->transportSpecific(f, "non"), 7));
     f->free(f);
 }
 
 // Tests get domain number parameter
-// i16 domainNumber(ptpmgmt_cfg _this, const char *section)
+// uint8_t domainNumber(ptpmgmt_cfg _this, const char *section)
 Test(ConfigFileTest, MethodDomainNumber)
 {
     ptpmgmt_cfg f = ptpmgmt_cfg_alloc();
     cr_expect(not(zero(ptr, f)));
     cr_expect(f->read_cfg(f, "utest/testing.cfg"));
-    cr_expect(eq(i16, f->domainNumber(f, "dumm"), 3));
-    cr_expect(eq(i16, f->domainNumber(f, "non"), 5));
+    cr_expect(eq(i8, f->domainNumber(f, "dumm"), 3));
+    cr_expect(eq(i8, f->domainNumber(f, "non"), 5));
     f->free(f);
 }
 
 // Tests get UDP IPv6 scope parameter
-// i16 udp6_scope(ptpmgmt_cfg _this, const char *section)
+// uint8_t udp6_scope(ptpmgmt_cfg _this, const char *section)
 Test(ConfigFileTest, MethodUdp6_scope)
 {
     ptpmgmt_cfg f = ptpmgmt_cfg_alloc();
     cr_expect(not(zero(ptr, f)));
     cr_expect(f->read_cfg(f, "utest/testing.cfg"));
-    cr_expect(eq(i16, f->udp6_scope(f, "dumm"), 0xf));
-    cr_expect(eq(i16, f->udp6_scope(f, "non"), 0xd));
+    cr_expect(eq(i8, f->udp6_scope(f, "dumm"), 0xf));
+    cr_expect(eq(i8, f->udp6_scope(f, "non"), 0xd));
     f->free(f);
 }
 
 // Tests get UDP IPv4 Time to live parameter
-// i16 udp_ttl(ptpmgmt_cfg _this, const char *section)
+// uint8_t udp_ttl(ptpmgmt_cfg _this, const char *section)
 Test(ConfigFileTest, MethodUdp_ttl)
 {
     ptpmgmt_cfg f = ptpmgmt_cfg_alloc();
     cr_expect(not(zero(ptr, f)));
     cr_expect(f->read_cfg(f, "utest/testing.cfg"));
-    cr_expect(eq(i16, f->udp_ttl(f, "dumm"), 7));
-    cr_expect(eq(i16, f->udp_ttl(f, "non"), 3));
+    cr_expect(eq(i8, f->udp_ttl(f, "dumm"), 7));
+    cr_expect(eq(i8, f->udp_ttl(f, "non"), 3));
     f->free(f);
 }
 
 // Tests get socket priority parameter
-// i16 socket_priority(ptpmgmt_cfg _this, const char *section)
+// uint8_t socket_priority(ptpmgmt_cfg _this, const char *section)
 Test(ConfigFileTest, MethodSocketPriority)
 {
     ptpmgmt_cfg f = ptpmgmt_cfg_alloc();
     cr_expect(not(zero(ptr, f)));
     cr_expect(f->read_cfg(f, "utest/testing.cfg"));
-    cr_expect(eq(i16, f->socket_priority(f, "dumm"), 7));
-    cr_expect(eq(i16, f->socket_priority(f, "non"), 11));
+    cr_expect(eq(i8, f->socket_priority(f, "dumm"), 7));
+    cr_expect(eq(i8, f->socket_priority(f, "non"), 11));
     f->free(f);
 }
 
 // Tests get network transport type parameter
-// i16 network_transport(ptpmgmt_cfg _this, const char *section)
+// uint8_t network_transport(ptpmgmt_cfg _this, const char *section)
 Test(ConfigFileTest, MethodNetworkTransport)
 {
     ptpmgmt_cfg f = ptpmgmt_cfg_alloc();
     cr_expect(not(zero(ptr, f)));
     cr_expect(f->read_cfg(f, "utest/testing.cfg"));
-    cr_expect(eq(i16, f->network_transport(f, "dumm"), '4'));
-    cr_expect(eq(i16, f->network_transport(f, "non"), '6'));
+    cr_expect(eq(i8, f->network_transport(f, "dumm"), '4'));
+    cr_expect(eq(i8, f->network_transport(f, "non"), '6'));
     f->free(f);
 }
